distinguish between models and dtos
