+++ src/auth/dtos/signInPayload.dto.ts
@@ -1,0 +1,15 @@
+import { IsNotEmpty, IsString } from "class-validator";
+
+export class SignInPayloadDto {
+    @IsString()
+    @IsNotEmpty()
+    username: string;
+
+    @IsString()
+    @IsNotEmpty()
+    password: string;
+
+    constructor(partial?: Partial<SignInPayloadDto>) {
+        Object.assign(this, partial);
+    }
+}
