+++ src/auth/dtos/signToken.dto.ts
@@ -1,0 +1,10 @@
+import { ApiProperty } from "@nestjs/swagger";
+
+export class SignTokenDto {
+    @ApiProperty({ type: String })
+    access_token: string;
+
+    constructor(partial?: Partial<SignTokenDto>) {
+        Object.assign(this, partial);
+    }
+}
