+++ src/comments/controllers/comments.controller.spec.ts
@@ -1,18 +1,18 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { CommentsController } from './comments.controller';
+import { Test, TestingModule } from "@nestjs/testing";
+import { CommentsController } from "./comments.controller";
 
-describe('CommentsController', () => {
-  let controller: CommentsController;
+describe("CommentsController", () => {
+    let controller: CommentsController;
 
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [CommentsController],
-    }).compile();
+    beforeEach(async () => {
+        const module: TestingModule = await Test.createTestingModule({
+            controllers: [CommentsController],
+        }).compile();
 
-    controller = module.get<CommentsController>(CommentsController);
-  });
+        controller = module.get<CommentsController>(CommentsController);
+    });
 
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
+    it("should be defined", () => {
+        expect(controller).toBeDefined();
+    });
 });
