+++ src/comments/dtos/hateSpeechResponse.dto.ts
