+++ src/comments/dtos/index.ts
@@ -1,0 +1,4 @@
+export { CommentCreationPayloadDto } from "./commentCreationPayload.dto";
+export { HateSpeechRequestPayloadDto } from "./hateSpeechRequestPayload.dto";
+export { HateSpeechResponseDto } from "./hateSpeechResponse.dto";
+export { VoteCommentPayloadDto, VoteType } from "./voteCommentPayload.dto";
