+++ src/comments/models/index.ts
@@ -1,6 +1,1 @@
 export { Comment } from "./comment";
-export { CommentCreationPayloadDto } from "./commentCreationPayload.dto";
-export { HateSpeechRequestPayloadDto } from "./hateSpeechRequestPayload.dto";
-export { HateSpeechResponseDto } from "./hateSpeechResponse.dto";
-export { VoteCommentPayloadDto, VoteType } from "./voteCommentPayload.dto";
-
