+++ src/comments/services/comments/comments.service.interface.ts
@@ -1,5 +1,5 @@
-import { Comment, VoteCommentPayloadDto } from "../../models";
-import { CommentCreationPayloadDto } from "../../models/commentCreationPayload.dto";
+import { Comment } from "../../models";
+import { VoteCommentPayloadDto, CommentCreationPayloadDto } from "../../dtos";
 
 export interface ICommentsService {
     authorNewComment(commentPayload: CommentCreationPayloadDto): Promise<Comment>;
