+++ src/comments/services/comments/comments.service.spec.ts
@@ -1,18 +1,18 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { CommentsService } from './comments.service';
+import { Test, TestingModule } from "@nestjs/testing";
+import { CommentsService } from "./comments.service";
 
-describe('CommentsService', () => {
-  let service: CommentsService;
+describe("CommentsService", () => {
+    let service: CommentsService;
 
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [CommentsService],
-    }).compile();
+    beforeEach(async () => {
+        const module: TestingModule = await Test.createTestingModule({
+            providers: [CommentsService],
+        }).compile();
 
-    service = module.get<CommentsService>(CommentsService);
-  });
+        service = module.get<CommentsService>(CommentsService);
+    });
 
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
+    it("should be defined", () => {
+        expect(service).toBeDefined();
+    });
 });
