+++ src/posts/dtos/hateSpeechResponse.dto.ts
