+++ src/posts/dtos/index.ts
@@ -1,0 +1,6 @@
+export { HateSpeechResponseDto } from "./hateSpeechResponse.dto";
+export { HateSpeechRequestPayloadDto } from "./hateSpeechRequestPayload.dto";
+export { DeletePostPayloadDto } from "./deletePostPayload.dto";
+export { VotePostPayloadDto, VoteType } from "./votePostPayload.dto";
+export { ReportPostPayloadDto } from "./reportPostPayload.dto";
+export { PostCreationPayloadDto } from "./postCreationPayload.dto";
