+++ src/posts/dtos/postCreationPayload.dto.ts
