+++ src/posts/dtos/votePostPayload.dto.ts
