+++ src/posts/models/index.ts
@@ -2,8 +2,3 @@ export { Post } from "./post";
 export { PostType } from "./postType";
 export { PostTag } from "./postTag";
 export { Award } from "./award";
-export { HateSpeechResponseDto } from "./hateSpeechResponse.dto";
-export { HateSpeechRequestPayloadDto } from "./hateSpeechRequestPayload.dto";
-export { DeletePostPayloadDto } from "./deletePostPayload.dto";
-export { VotePostPayloadDto, VoteType } from "./votePostPayload.dto";
-export { ReportPostPayloadDto } from "./reportPostPayload.dto";
