+++ src/users/models/index.ts
@@ -1,4 +1,3 @@
-export { RegisterUserPayloadDto } from "./registerUserPayload.dto";
 export { User } from "./user";
 export { Role } from "./role";
 export { Gender } from "./gender";
